Exported TradeOffer class in steam-tradeoffer-manager
